# src/main/java/com/letrasypapeles/backend/dto/BranchDTO.java
package com.letrasypapeles.backend.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor

public class BranchDTO {
  private String name;
	private String address;
	private String region;
}