# src/main/java/com/letrasypapeles/backend/dto/CartItemRequest.java
package com.letrasypapeles.backend.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@Builder
@AllArgsConstructor
public class CartItemRequest {
  private Long productId;
  private Integer quantity;
}