# src/main/java/com/letrasypapeles/backend/dto/CartRequest.java
package com.letrasypapeles.backend.dto;

import java.util.List;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@Builder
@AllArgsConstructor

public class CartRequest {
  private Long clientId;
  private List<CartItemRequest> items;
}