# src/main/java/com/letrasypapeles/backend/dto/ProductDTO.java
package com.letrasypapeles.backend.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor

public class ProductDTO {
	private String name;
	private String details;
	private double price;
	private int stock;
}