# src/main/java/com/letrasypapeles/backend/entity/Order.java
package com.letrasypapeles.backend.entity;

import jakarta.persistence.*;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.LocalDateTime;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder

@Table(name ="tbl_orders")
@Entity
public class Order {
	@Id
	@GeneratedValue(strategy = GenerationType.IDENTITY)
	private Long id;

	private String orderNumber;
  private LocalDateTime issueDate;
  private Double totalAmount;

	@OneToOne
  @JoinColumn(name = "cart_id", unique = true)
  private Cart cart;

}