# src/main/resources/api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Letras y Papeles API",
    "description": "Endpoints para la implementación del proyecto Letras y Papeles",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/order/{id}": {
      "get": {
        "tags": [
          " ** Endpoints Orden **"
        ],
        "summary": "Obtiene una orden por su ID",
        "description": "Este endpoint permite obtener una orden específica utilizando su ID",
        "operationId": "obtenerPorId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador único de la orden",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Orden obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "404": {
            "description": "Orden no encontrada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelOrder"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelOrder"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          " ** Endpoints Orden **"
        ],
        "summary": "Actualiza una orden existente",
        "description": "Este endpoint permite actualizar los detalles de una orden existente",
        "operationId": "actualizarOrden",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador único de la orden",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "description": "Datos de la orden a actualizar",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Order"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Orden actualizada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "404": {
            "description": "Orden no encontrada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelOrder"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelOrder"
                }
              }
            }
          }
        }
      }
    },
    "/api/branch/{id}": {
      "get": {
        "tags": [
          "Sucursales"
        ],
        "summary": "Obtiene una sucursal por su ID",
        "description": "Recupera los detalles de una sucursal específica utilizando su identificador único",
        "operationId": "obtenerPorId_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador único de la sucursal",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Sucursal encontrada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Branch"
                }
              }
            }
          },
          "404": {
            "description": "Sucursal no encontrada con el ID proporcionado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Branch"
                }
              }
            }
          },
          "400": {
            "description": "ID de sucursal inválido.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Branch"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Sucursales"
        ],
        "summary": "Actualiza una sucursal existente",
        "description": "Permite modificar completamente los datos de una sucursal existente identificándola por su ID",
        "operationId": "actualizarSucursal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador único de la sucursal a actualizar",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "description": "Datos actualizados de la sucursal",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Branch"
              },
              "examples": {
                "Sucursal actualizada": {
                  "summary": "Ejemplo de actualización de sucursal",
                  "description": "Sucursal actualizada",
                  "value": {
                    "name": "Sucursal Test Actuliazada",
                    "address": "Av. Test 123",
                    "city": "Ciudad de Test"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Sucursal actualizada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Branch"
                }
              }
            }
          },
          "404": {
            "description": "Sucursal no encontrada con el ID proporcionado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Branch"
                }
              }
            }
          },
          "400": {
            "description": "Datos de la sucursal inválidos o incompletos",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Branch"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Sucursales"
        ],
        "summary": "Elimina una sucursal",
        "description": "Permite eliminar una sucursal específica del sistema utilizando su identificador único",
        "operationId": "delete_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador único de la sucursal a eliminar",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Sucursal eliminada exitosamente",
            "content": {
              "application/json": {
                "example": {
                  "message": "Sucursal eliminada exitosamente"
                }
              }
            }
          },
          "404": {
            "description": "Sucursal no encontrada con el ID proporcionado",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/carts": {
      "post": {
        "tags": [
          "Cart"
        ],
        "operationId": "processCart",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/branch": {
      "get": {
        "tags": [
          "Sucursales"
        ],
        "summary": "Obtiene una lista de todas las sucursales",
        "description": "Este endpoint devuelve una lista completa de todas las sucursales registradas en el sistema",
        "operationId": "obtenerTodas",
        "responses": {
          "200": {
            "description": "Lista de sucursales obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Branch"
                  }
                }
              }
            }
          },
          "204": {
            "description": "No hay sucursales disponibles (La lista está vacía)",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Branch"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Branch"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Sucursales"
        ],
        "summary": "Crea una nueva sucursal",
        "description": "Permite registrar una nueva sucursal en el sistema con todos los datos necesarios",
        "operationId": "crearSucursal",
        "requestBody": {
          "description": "Datos de la nueva sucursal a crear",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BranchDTO"
              },
              "examples": {
                "Sucursal ejemplo": {
                  "summary": "Ejemplo de una sucursal completa",
                  "description": "Sucursal ejemplo",
                  "value": {
                    "name": "Sucursal Test Crear",
                    "address": "Av. Test 123, Col. Centro",
                    "region": "Región Test"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Sucursal creada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Branch"
                }
              }
            }
          },
          "400": {
            "description": "Datos de la sucursal inválidos o incompletos",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Branch"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto: una sucursal con datos similares ya existe",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Branch"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/carts/{id}": {
      "get": {
        "tags": [
          "Cart"
        ],
        "operationId": "getCartById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CartResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/carts/orders/{id}": {
      "get": {
        "tags": [
          "Cart"
        ],
        "operationId": "getOrderById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Cart": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "client": {
            "$ref": "#/components/schemas/Client"
          },
          "cartItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartItem"
            }
          }
        }
      },
      "CartItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          }
        }
      },
      "Client": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "fidelityPoints": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "orderNumber": {
            "type": "string"
          },
          "issueDate": {
            "type": "string",
            "format": "date-time"
          },
          "totalAmount": {
            "type": "number",
            "format": "double"
          },
          "client": {
            "$ref": "#/components/schemas/Client"
          },
          "cart": {
            "$ref": "#/components/schemas/Cart"
          }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "details": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "sku": {
            "type": "string"
          }
        }
      },
      "OrderResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "orderNumber": {
            "type": "string"
          },
          "issueDate": {
            "type": "string",
            "format": "date-time"
          },
          "totalAmount": {
            "type": "number",
            "format": "double"
          },
          "cartId": {
            "type": "integer",
            "format": "int64"
          },
          "clientId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "EntityModelOrder": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "orderNumber": {
            "type": "string"
          },
          "issueDate": {
            "type": "string",
            "format": "date-time"
          },
          "totalAmount": {
            "type": "number",
            "format": "double"
          },
          "client": {
            "$ref": "#/components/schemas/Client"
          },
          "cart": {
            "$ref": "#/components/schemas/Cart"
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
      "Branch": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "region": {
            "type": "string"
          }
        }
      },
...
      "CartItemRequest": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "format": "int64"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "CartRequest": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "integer",
            "format": "int64"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartItemRequest"
            }
          }
        }
      },
      "BranchDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "region": {
            "type": "string"
          }
        }
      },
...
      "CartItemResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "productId": {
            "type": "integer",
            "format": "int64"
          },
          "productName": {
            "type": "string"
          },
          "productSku": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "CartResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "cartItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartItemResponse"
            }
          },
          "orderId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "Link": {
        "type": "object",
        "properties": {
          "href": {
            "type": "string"
          },
          "hreflang": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "deprecation": {
            "type": "string"
          },
          "profile": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "templated": {
            "type": "boolean"
          }
        }
      },
      "Links": {
        "type": "object",
        "additionalProperties": {
          "$ref": "#/components/schemas/Link"
        }
      }
    }
  }
}